refactor(tests): helpers.go -> accessors.go
